Move indexed classes to a separate module, add itoList
