+++ optics-core/src/Optics/Internal/IxTraversal.hs
@@ -9,7 +9,7 @@ module Optics.Internal.IxTraversal where
 import Data.Profunctor.Indexed
 
 import Optics.Internal.Fold
-import Optics.Internal.Indexed
+import Optics.Internal.Indexed.Classes
 import Optics.Internal.IxFold
 import Optics.Internal.IxSetter
 import Optics.Internal.Optic
