+++ optics-core/src/Optics/IxTraversal.hs
@@ -82,6 +82,7 @@ import Data.Functor.Identity
 import Data.Profunctor.Indexed
 
 import Optics.Internal.Indexed
+import Optics.Internal.Indexed.Classes
 import Optics.Internal.IxTraversal
 import Optics.Internal.Optic
 import Optics.Internal.Utils
